+++ test_opensearchpy/test_connection/__init__.py
@@ -23,23 +23,3 @@
 #  KIND, either express or implied.  See the License for the
 #  specific language governing permissions and limitations
 #  under the License.
-
-import sys
-
-from .._async.helpers.actions import async_bulk as async_bulk
-from .._async.helpers.actions import async_reindex as async_reindex
-from .._async.helpers.actions import async_scan as async_scan
-from .._async.helpers.actions import async_streaming_bulk as async_streaming_bulk
-from .actions import _chunk_actions as _chunk_actions
-from .actions import _process_bulk_chunk as _process_bulk_chunk
-from .actions import bulk as bulk
-from .actions import expand_action as expand_action
-from .actions import parallel_bulk as parallel_bulk
-from .actions import reindex as reindex
-from .actions import scan as scan
-from .actions import streaming_bulk as streaming_bulk
-from .asyncsigner import AWSV4SignerAsyncAuth as AWSV4SignerAsyncAuth
-from .errors import BulkIndexError as BulkIndexError
-from .errors import ScanError as ScanError
-from .signer import AWSV4SignerAuth as AWSV4SignerAuth
-from .signer import RequestsAWSV4SignerAuth, Urllib3AWSV4SignerAuth